pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sem-mem"
version = "0.1.0"
description = "Tiered semantic memory for AI agents with HNSW-based retrieval"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "David" }
]
keywords = ["ai", "memory", "rag", "semantic-search", "chatbot", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pypdf>=3.17.0",
    "python-dotenv>=1.0.0",
    "hnswlib>=0.8.0",
]

[project.optional-dependencies]
app = [
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "plotly>=5.18.0",
]
server = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]
all = [
    "sem-mem[app,server,dev]",
]

[project.urls]
Homepage = "https://github.com/david/sem-mem"
Repository = "https://github.com/david/sem-mem"

[tool.hatch.build.targets.wheel]
packages = ["sem_mem"]

[tool.ruff]
line-length = 100
target-version = "py310"
